修改竞猜排行榜用户头像不显示问题； 优化每日任务转盘表达； 抽取静态常量到常量类；
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@647101 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/rank_info_entity.dart
@@ -6,7 +6,7 @@ export 'package:arm_chair_quaterback/generated/json/rank_info_entity.g.dart';
 @JsonSerializable()
 class RankInfoEntity {
   late double guessCount;
-  late double teamLogo;
+  late int teamLogo;
   late String teamName;
   late double chip;
   late double createTime;
